+++ lib/recording_studio_notifications/engine.rb
@@ -25,6 +25,20 @@ module RecordingStudioNotifications
                          RecordingStudioNotifications.configuration.hooks.controller_extensions_for(extension_keys_for(target)))
       end
 
+      def prepend_admin_table_wrap_views!
+        return unless defined?(RecordingStudioAdmin::ApplicationController)
+
+        view_path = root.join("app/views").to_s
+        controller = RecordingStudioAdmin::ApplicationController
+        existing_paths = controller.view_paths.map(&:to_s)
+        controller.prepend_view_path(view_path) unless existing_paths.include?(view_path)
+
+        helper_module = RecordingStudioNotifications::Admin::TableWrapHelper
+        return if controller._helpers.included_modules.include?(helper_module)
+
+        controller.helper(helper_module)
+      end
+
       private
 
       def define_default_accessible_action(action, &)
@@ -151,10 +165,12 @@ module RecordingStudioNotifications
 
       require "recording_studio_notifications/admin/all_notifications_screen"
       require "recording_studio_notifications/admin/all_notifications_section"
+      require "recording_studio_notifications/admin/table_wrap_helper"
 
       config.to_prepare do
         RecordingStudioAdmin.register_screen(RecordingStudioNotifications::Admin::AllNotificationsScreen)
         RecordingStudioAdmin.register_section(RecordingStudioNotifications::Admin::AllNotificationsSection)
+        RecordingStudioNotifications::Engine.prepend_admin_table_wrap_views!
       end
     end